requirements.txt
aiogram==2.10.1
aiohttp==3.7.2
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
idna==2.10
loguru==0.5.3
lxml==4.6.1
multidict==5.0.0
pafy==0.5.5
Pillow==8.0.1
pkg-resources==0.0.0
pytz==2020.4
requests==2.24.0
schedule==0.6.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
youtube-dl==2021.1.16
